3rd
Added Button and door. The button changes color when the player is on it and with that the door opens. when the player gets down from the button the door closes. so the player has to stay on the button to keep it open. The color change in the button is instant